Fix drac implementation of set_boot_device
The drac implementation of set_boot_device was not recognizing when
it did not actually need to change the boot order, which caused it to
try running a config job when there was nothing to do.

This adds a check for that condition, and exits early if there is
nothing to do.

Closes-Bug: 1454259
Change-Id: I48971c79ba3d745056ff3d3157fde08b3d566f94