DRAC: support for RAID via vendor_passthru
The patch adds support for RAID configuration via the vendor_passthru
interface. The following low level operations are supported:
- listing RAID controllers
- listing virtual disks
- listing physical disks
- creating a single virtual disk
- deleting a single virtual disk
- applying pending RAID configuration
- deleting pending RAID configuration

The methods in ironic.drivers.modules.drac.raid are implemented in a way
that they will also be used by the generic RAID interface. The benefit
of exposing them via vendor_passthru is that the user have more
flexibility for RAID configuration.

The patch also provides way to check the status of the configuration job
created on the DRAC card, since the changes are applied asynchronously.
With these methods the user can follow the progress of the RAID
configuration.

Change-Id: Ic2e255ea2bccc48f41e5f8368ac2c5b6eb345950